Drag & Drop Update + SCSS
I've moved the item list from it's own component to combine it with the cards component. This means only managing a single state and allows for a drag and drop list re-ordering while allowing you to still add and delete from the list.

Included the use of Sass however, you dont strictly need it. It can be removed. Only included it as it was required as part of trying to get the drag and drop to work.